requirements.txt
# Core Dependencies
selenium>=4.15.0
beautifulsoup4>=4.12.0
webdriver-manager>=4.0.0
requests>=2.31.0

# Database
pymongo>=4.6.0

# AI/LLM (let CrewAI manage its dependencies)
google-genai>=1.0.0
crewai>=0.80.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.9
langchain-text-splitters>=0.0.1
pypdf>=4.0.0
chromadb>=0.4.22

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Frontend
flask>=3.0.0
flask-cors>=4.0.0

# PDF Generation
reportlab>=4.0.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
pillow>=10.1.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.12.0
flake8>=6.1.0
pylint>=3.0.0
bandit>=1.7.0

litellm>=1.40.0
groq>=0.9.0